+++ R/geom_protein.R
@@ -29,7 +29,7 @@
 #' @importFrom stringr str_locate
 #' @importFrom GenomicRanges reduce GRanges setdiff
 #' @importFrom IRanges IRanges
-#' @importFrom ggplot2 ggplot geom_rect geom_text aes aes_string scale_x_continuous
+#' @importFrom ggplot2 ggplot geom_rect geom_text aes aes_string
 #' @importFrom ggpp annotate
 #' @importFrom scales scientific
 #' @export
@@ -42,129 +42,154 @@
 #' # protein.id = "sp|
 #' # ggplot() +
 #' #     geom_peptide(coverage.file = coverage.file, fasta.file = fasta.file, protein.id = protein.id)
-geom_protein = function(coverage.file, fasta.file, protein.id, XCorr.threshold = 2,
-                        confidence = "High", contaminant = NULL, remove.na = TRUE,
-                        color = "grey", mark.bare = TRUE, mark.color = "red", mark.alpha = 0.5,
-                        show.table = TRUE, table.position = c("right_top", "left_top", "left_bottom", "right_bottom"),
-                        table.size = 4, table.color = "black", range.size = 3, range.position = c("in", "out")){
+geom_protein <- function(coverage.file, fasta.file, protein.id, XCorr.threshold = 2,
+                         confidence = "High", contaminant = NULL, remove.na = TRUE,
+                         color = "grey", mark.bare = TRUE, mark.color = "red", mark.alpha = 0.5,
+                         show.table = TRUE, table.position = c("right_top", "left_top", "left_bottom", "right_bottom"),
+                         table.size = 4, table.color = "black", range.size = 3, range.position = c("in", "out")) {
   # check parameters
   table.position <- match.arg(arg = table.position)
   range.position <- match.arg(arg = range.position)
 
   # load coverage dataframe
-  coverage.df = openxlsx::read.xlsx(coverage.file)
+  coverage.df <- openxlsx::read.xlsx(coverage.file)
   # remove suffix and prefix string
-  coverage.df$Annotated.Sequence = gsub(pattern = ".*\\.(.*)\\..*", replacement = "\\1", x = coverage.df$Annotated.Sequence)
+  coverage.df$Annotated.Sequence <- gsub(pattern = ".*\\.(.*)\\..*", replacement = "\\1", x = coverage.df$Annotated.Sequence)
   # filter converge according to confidence
-  if(!is.null(confidence)){
-    coverage.df = coverage.df[coverage.df[, "Confidence"] == confidence, ]
+  if (!is.null(confidence)) {
+    coverage.df <- coverage.df[coverage.df[, "Confidence"] == confidence, ]
   }
   # filter converge according to contaminant
-  if(!is.null(contaminant)){
-    coverage.df = coverage.df[coverage.df[, "Contaminant"] == contaminant, ]
+  if (!is.null(contaminant)) {
+    coverage.df <- coverage.df[coverage.df[, "Contaminant"] == contaminant, ]
   }
   # filter converge according to cross-correlation
-  if(!is.null(XCorr.threshold)){
-    xcorr.index = grep(pattern = "XCorr", x = colnames(coverage.df))
-    coverage.df = coverage.df[coverage.df[, xcorr.index] >= XCorr.threshold, ]
+  if (!is.null(XCorr.threshold)) {
+    xcorr.index <- grep(pattern = "XCorr", x = colnames(coverage.df))
+    coverage.df <- coverage.df[coverage.df[, xcorr.index] >= XCorr.threshold, ]
   }
   # get abundance cols
-  abundance.col = grep(pattern = "Abundance", x = colnames(coverage.df), value = TRUE)
+  abundance.col <- grep(pattern = "Abundance", x = colnames(coverage.df), value = TRUE)
   # remove na abundance
-  if(remove.na){
-    coverage.df = coverage.df %>% dplyr::filter(!is.na(.data[[abundance.col]]))
+  if (remove.na) {
+    coverage.df <- coverage.df %>% dplyr::filter(!is.na(.data[[abundance.col]]))
   }
   # sum abundance of duplicated Annotated.Sequence
-  coverage.df = coverage.df %>%
+  coverage.df <- coverage.df %>%
     dplyr::group_by(.data[["Annotated.Sequence"]]) %>%
     dplyr::summarise(Abundance = sum(.data[[abundance.col]])) %>%
     as.data.frame()
-  colnames(coverage.df) = c("peptide", "abundance")
+  colnames(coverage.df) <- c("peptide", "abundance")
   # check the coverage dataframe
-  if(nrow(coverage.df) == 0){
+  if (nrow(coverage.df) == 0) {
     stop("There is no valid peptide, please check!")
   }
 
   # load genome fasta
-  aa.set = Biostrings::readAAStringSet(fasta.file)
-  protein.index = which(names(aa.set) == protein.id)
-  if(length(protein.index) == 1){
-    aa.set.used = aa.set[protein.index]
-    aa.seq.used = paste(aa.set.used)
-  }else if(length(protein.index) > 1){
+  aa.set <- Biostrings::readAAStringSet(fasta.file)
+  protein.index <- which(names(aa.set) == protein.id)
+  if (length(protein.index) == 1) {
+    aa.set.used <- aa.set[protein.index]
+    aa.seq.used <- paste(aa.set.used)
+  } else if (length(protein.index) > 1) {
     stop("Please check the protein.id you provided, there is more than one in provided fasta file!")
-  }else{
+  } else {
     stop("Please check the protein.id you provided, it can't be found in provided fasta file!")
   }
 
   # get the region
-  aa.anno.region = sapply(coverage.df$peptide, function(x){
-    stringr::str_locate(pattern =x, aa.seq.used)
-  }) %>% t() %>% as.data.frame()
-  colnames(aa.anno.region) = c("start", "end")
+  aa.anno.region <- sapply(coverage.df$peptide, function(x) {
+    stringr::str_locate(pattern = x, aa.seq.used)
+  }) %>%
+    t() %>%
+    as.data.frame()
+  colnames(aa.anno.region) <- c("start", "end")
 
   # merge
-  coverage.final = merge(coverage.df, aa.anno.region, by.x = "peptide", by.y = 0, all.x = TRUE)
-  coverage.final = coverage.final %>% dplyr::arrange(.data[["start"]], .data[["end"]])
+  coverage.final <- merge(coverage.df, aa.anno.region, by.x = "peptide", by.y = 0, all.x = TRUE)
+  coverage.final <- coverage.final %>% dplyr::arrange(.data[["start"]], .data[["end"]])
+  coverage.final$ProteinID <- protein.id
 
   # get coverage positions
-  coverage.pos =
+  coverage.pos <-
     GenomicRanges::reduce(GenomicRanges::GRanges(protein.id, IRanges::IRanges(coverage.final$start, coverage.final$end))) %>%
     as.data.frame()
-  coverage.pos$strand = NULL
-  colnames(coverage.pos) = c("ProteinID", "start", "end", "width")
-  coverage.pos$Type = "covered"
+  coverage.pos$strand <- NULL
+  colnames(coverage.pos) <- c("ProteinID", "start", "end", "width")
+  coverage.pos$Type <- "covered"
   # get coverage rate
-  coverage.rate = round(sum(coverage.pos$width)*100/nchar(aa.seq.used), 2)
+  coverage.rate <- round(sum(coverage.pos$width) * 100 / nchar(aa.seq.used), 2)
   # non-cover position
-  non.coverage.pos =
-    GenomicRanges::setdiff(GenomicRanges::GRanges(protein.id, IRanges::IRanges(1, nchar(aa.seq.used))),
-                           GenomicRanges::GRanges(protein.id, IRanges::IRanges(coverage.final$start, coverage.final$end))) %>%
+  non.coverage.pos <-
+    GenomicRanges::setdiff(
+      GenomicRanges::GRanges(protein.id, IRanges::IRanges(1, nchar(aa.seq.used))),
+      GenomicRanges::GRanges(protein.id, IRanges::IRanges(coverage.final$start, coverage.final$end))
+    ) %>%
     as.data.frame()
-  non.coverage.pos$strand = NULL
-  colnames(non.coverage.pos) = c("ProteinID", "start", "end", "width")
-  non.coverage.pos$Type = "bare"
+  non.coverage.pos$strand <- NULL
+  colnames(non.coverage.pos) <- c("ProteinID", "start", "end", "width")
+  non.coverage.pos$Type <- "bare"
   # coverage summary
-  coverage.summary = rbind(coverage.pos, non.coverage.pos) %>% as.data.frame()
+  coverage.summary <- rbind(coverage.pos, non.coverage.pos) %>% as.data.frame()
+
+  # create whole range
+  non.coverage.final <- data.frame(
+    ProteinID = protein.id, peptide = "Empty", abundance = 0,
+    start = non.coverage.pos$start, end = non.coverage.pos$end
+  )
+  coverage.final <- coverage.final[c("ProteinID", "peptide", "abundance", "start", "end")]
+  coverage.final <- rbind(coverage.final, non.coverage.final) %>%
+    as.data.frame() %>%
+    dplyr::arrange(.data[["start"]], .data[["end"]])
 
   # coverage rect
-  coverage.rect = geom_rect(data = coverage.final, mapping = aes_string(xmin = "start", xmax = "end",
-                                                                        ymin = "0", ymax = "abundance"),
-                            show.legend = FALSE, fill = color)
+  coverage.rect <- geom_rect(
+    data = coverage.final, mapping = aes_string(
+      xmin = "start", xmax = "end",
+      ymin = "0", ymax = "abundance"
+    ),
+    show.legend = FALSE, fill = color
+  )
   plot.ele <- list(coverage.rect)
   # mark bare
-  if(mark.bare){
-    bare.rect = geom_rect(data = non.coverage.pos, mapping = aes_string(xmin = "start", xmax = "end",
-                                                                        ymin = "0", ymax = "Inf"),
-                          show.legend = F, fill = mark.color, alpha = mark.alpha)
+  if (mark.bare) {
+    bare.rect <- geom_rect(
+      data = non.coverage.pos, mapping = aes_string(
+        xmin = "start", xmax = "end",
+        ymin = "0", ymax = "Inf"
+      ),
+      show.legend = F, fill = mark.color, alpha = mark.alpha
+    )
     plot.ele <- append(plot.ele, bare.rect)
   }
   # summary table
-  if(show.table){
+  if (show.table) {
     # table position
-    if(table.position == "left_top"){
-      table.x = 0
-      table.y = max(coverage.final[ , "abundance"])
-    }else if(table.position == "right_top"){
-      table.x = nchar(aa.seq.used)
-      table.y = max(coverage.final[ , "abundance"])
-    }else if(table.position == "left_bottom"){
-      table.x = 0
-      table.y = 0
-    }else if(table.position == "right_bottom"){
-      table.x = nchar(aa.seq.used)
-      table.y = 0
+    if (table.position == "left_top") {
+      table.x <- 0
+      table.y <- max(coverage.final[, "abundance"])
+    } else if (table.position == "right_top") {
+      table.x <- nchar(aa.seq.used)
+      table.y <- max(coverage.final[, "abundance"])
+    } else if (table.position == "left_bottom") {
+      table.x <- 0
+      table.y <- 0
+    } else if (table.position == "right_bottom") {
+      table.x <- nchar(aa.seq.used)
+      table.y <- 0
     }
-    summary.table = ggpp::annotate(geom = "table", label = list(coverage.summary), x= table.x, y=table.y,
-                                   color = table.color, size = table.size)
+    summary.table <- ggpp::annotate(
+      geom = "table", label = list(coverage.summary), x = table.x, y = table.y,
+      color = table.color, size = table.size
+    )
     plot.ele <- append(plot.ele, summary.table)
   }
   # range position
   if (range.position == "in") {
     # prepare range
-    max.abundance = CeilingNumber(max(coverage.final$abundance))
-    abundance.range = data.frame(label = paste0("[0, ", scales::scientific(max.abundance, digits = 2), "]"))
-    range.text = geom_text(
+    max.abundance <- CeilingNumber(max(coverage.final$abundance))
+    abundance.range <- data.frame(label = paste0("[0, ", scales::scientific(max.abundance, digits = 2), "]"))
+    range.text <- geom_text(
       data = abundance.range,
       mapping = aes(x = -Inf, y = Inf, label = label),
       hjust = 0,
@@ -173,7 +198,5 @@ geom_protein = function(coverage.file, fasta.file, protein.id, XCorr.threshold =
     )
     plot.ele <- append(plot.ele, range.text)
   }
-  # change x scale
-  plot.ele <- append(plot.ele, scale_x_continuous(limits = c(1, nchar(aa.seq.used)), expand = c(0, 0)))
   return(plot.ele)
 }